Complete render of homepage
